+++ core/src/main/java/com/novoda/noplayer/subtitles/SubtitlesStyle.java
@@ -1,0 +1,11 @@
+package com.novoda.noplayer.subtitles;
+
+public interface SubtitlesStyle {
+
+    int backgroundColorOr(int fallbackColor);
+
+    int foregroundColorOr(int fallbackColor);
+
+    float scaleTextSize(float textSize);
+
+}
