+++ core/src/main/java/com/novoda/noplayer/subtitles/SystemCaptionPreferences.java
@@ -1,14 +1,7 @@
 package com.novoda.noplayer.subtitles;
 
 import android.content.Context;
-import android.os.Build.VERSION;
 import android.support.annotation.NonNull;
-import android.support.annotation.RequiresApi;
-import android.view.accessibility.CaptioningManager;
-
-import static android.os.Build.VERSION_CODES.KITKAT;
-import static android.view.accessibility.CaptioningManager.CaptionStyle;
-import static java.util.Objects.requireNonNull;
 
 public class SystemCaptionPreferences {
 
@@ -18,82 +11,15 @@ public class SystemCaptionPreferences {
         this.context = context;
     }
 
-    public Style getStyle() {
+    @NonNull
+    public SubtitlesStyle getStyle() {
         return getStyle(context);
     }
 
-    public interface Style {
-
-        int backgroundColorOr(int fallbackColor);
-
-        int foregroundColorOr(int fallbackColor);
-
-        float scaleTextSize(float textSize);
-    }
-
-    private static Style getStyle(Context context) {
-        if (VERSION.SDK_INT >= KITKAT) {
-            final CaptioningManager captioningManager = requireCaptionManager(context);
-            CaptionStyle systemStyle = captioningManager.getUserStyle();
-            float fontScale = captioningManager.getFontScale();
-            return new KitkatStyle(systemStyle, fontScale);
-        } else {
-            return NO_STYLE;
-        }
-    }
-
-    @RequiresApi(api = KITKAT)
     @NonNull
-    private static CaptioningManager requireCaptionManager(Context context) {
-        Object service = context.getSystemService(Context.CAPTIONING_SERVICE);
-        return requireNonNull((CaptioningManager) service);
+    private static SubtitlesStyle getStyle(Context context) {
+        return SubtitlesStyleFactory.create(context);
     }
 
-    private static final Style NO_STYLE = new Style() {
-
-        @Override
-        public int backgroundColorOr(int fallbackColor) {
-            return fallbackColor;
-        }
-
-        @Override
-        public int foregroundColorOr(int fallbackColor) {
-            return fallbackColor;
-        }
-
-        @Override
-        public float scaleTextSize(float textSize) {
-            return textSize;
-        }
-
-    };
-
-    @RequiresApi(api = KITKAT)
-    private static class KitkatStyle implements Style {
-
-        private final CaptionStyle captionStyle;
-        private final float fontScale;
-
-        KitkatStyle(CaptionStyle captionStyle, float fontScale) {
-            this.captionStyle = captionStyle;
-            this.fontScale = fontScale;
-        }
-
-        @Override
-        public int backgroundColorOr(int fallbackColor) {
-            return captionStyle.backgroundColor;
-        }
-
-        @Override
-        public int foregroundColorOr(int fallbackColor) {
-            return captionStyle.foregroundColor;
-        }
-
-        @Override
-        public float scaleTextSize(float textSize) {
-            return textSize * fontScale;
-        }
-
-    }
 
 }
